Ch.3 Programming Project Problem 1
Write a program that produces images of Christmas trees as output. It should have a method with two parameters:
one for the number of segments in the tree and one for the height of each segment. For example, the tree shown here
on the left has three segments of height 4 and the one on the right has two segments of height 5: